* circular dependency workaround
