+++ src/Data/EntityEntry.ts
@@ -4,21 +4,14 @@ import { EntityState } from "./EntityState";
 import { IEntityEntryOption } from "./Interface/IEntityEntry";
 import { RelationEntry } from "./RelationEntry";
 import { IRelationMetaData } from "../MetaData/Interface/IRelationMetaData";
-import { EmbeddedRelationMetaData } from "../MetaData/EmbeddedColumnMetaData";
 import { EventHandlerFactory } from "../Event/EventHandlerFactory";
 import { IEventHandler } from "../Event/IEventHandler";
 import { propertyChangeHandlerMetaKey, propertyChangeDispatherMetaKey, relationChangeHandlerMetaKey, relationChangeDispatherMetaKey } from "../Decorator/DecoratorKey";
-import { EmbeddedEntityEntry } from "./EmbeddedEntityEntry";
 import { FunctionExpression } from "../ExpressionBuilder/Expression/FunctionExpression";
 import { MemberAccessExpression } from "../ExpressionBuilder/Expression/MemberAccessExpression";
 import { ParameterExpression } from "../ExpressionBuilder/Expression/ParameterExpression";
 import { IEntityMetaData } from "../MetaData/Interface/IEntityMetaData";
 import { ExpressionExecutor } from "../ExpressionBuilder/ExpressionExecutor";
-
-let embeddedEntityEntry: typeof EmbeddedEntityEntry;
-(async () => {
-    embeddedEntityEntry = (await import("./EmbeddedEntityEntry")).EmbeddedEntityEntry;
-})();
 
 export class EntityEntry<T = any> implements IEntityEntryOption<T> {
     private _state: EntityState;
@@ -122,53 +115,13 @@ export class EntityEntry<T = any> implements IEntityEntryOption<T> {
     public get isCompletelyLoaded() {
         return this.dbSet.metaData.columns.all(o => (this.entity as any)[o.propertyName] !== undefined);
     }
-    private originalValues: Map<keyof T, any> = new Map();
+    // TODO: private
+    public originalValues: Map<keyof T, any> = new Map();
     public isPropertyModified(prop: keyof T) {
         return this.originalValues.has(prop);
     }
     public getOriginalValue(prop: keyof T) {
         return this.originalValues.get(prop);
-    }
-    protected onPropertyChanged(entity: T, param: IChangeEventParam<T>) {
-        if (this.dbSet.primaryKeys.contains(param.column)) {
-            // primary key changed, update dbset entry dictionary.
-            const oldKey = this.key;
-            this.dbSet.updateEntryKey(this);
-
-            // TODO: cascade update issue
-            // update all relation refer to this entity.
-            for (const prop in this.relationMap) {
-                const relationGroup = this.relationMap[prop];
-                if (!relationGroup)
-                    continue;
-
-                for (const [, relation] of relationGroup) {
-                    const entry = relation.masterEntry === this ? relation.slaveEntry : relation.masterEntry;
-                    entry.updateRelationKey(relation, oldKey);
-                }
-            }
-        }
-
-        if (param.oldValue !== param.newValue && param.column instanceof EmbeddedRelationMetaData) {
-            const embeddedDbSet = this.dbSet.dbContext.set(param.column.target.type);
-            new embeddedEntityEntry(embeddedDbSet, param.newValue, this);
-        }
-
-        if (this.enableTrackChanges && (this.state === EntityState.Modified || this.state === EntityState.Unchanged) && param.oldValue !== param.newValue) {
-            const oriValue = this.originalValues.get(param.column.propertyName);
-            if (oriValue === param.newValue) {
-                this.originalValues.delete(param.column.propertyName);
-                if (this.originalValues.size <= 0) {
-                    this.state = EntityState.Unchanged;
-                }
-            }
-            else if (oriValue === undefined && param.oldValue !== undefined && !param.column.isReadOnly) {
-                this.originalValues.set(param.column.propertyName, param.oldValue);
-                if (this.state === EntityState.Unchanged) {
-                    this.state = EntityState.Modified;
-                }
-            }
-        }
     }
     protected onRelationChanged(entity: T, param: IRelationChangeEventParam) {
         for (const item of param.entities) {
@@ -225,7 +178,8 @@ export class EntityEntry<T = any> implements IEntityEntryOption<T> {
             relGroup.delete(key);
         }
     }
-    protected updateRelationKey(relationEntry: RelationEntry<T, any> | RelationEntry<any, T>, oldEntityKey: string) {
+    // TODO: protected
+    public updateRelationKey(relationEntry: RelationEntry<T, any> | RelationEntry<any, T>, oldEntityKey: string) {
         const oldKey = relationEntry.slaveRelation.fullName + ":" + oldEntityKey;
         this.relationMap[oldKey] = undefined;
         relationEntry.join();
@@ -417,3 +371,5 @@ export class EntityEntry<T = any> implements IEntityEntryOption<T> {
 
     //#endregion
 }
+
+import "./EntityEntry.partial";
