--- src/Extensions/index.ts
@@ -1,6 +1,0 @@
-import "./ArrayExtension";
-import "./ArrayItemTypeExtension";
-import "./DateExtension";
-import "./EnumerableExtension";
-import "./QueryableExtension";
-import "./StringExtension";
